# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flock Energy - Urja Meter Ops API",
    "description": "Clean REST API proxy layer over the legacy Urja Meter Ops portal. See /docs for interactive testing.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/meters": {
      "get": {
        "summary": "List Meters",
        "description": "List meters (paginated), mirroring the portal's meter list/search.",
        "operationId": "list_meters_api_v1_meters_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Search by meter number or serial",
              "default": "",
              "title": "Q"
            },
            "description": "Search by meter number or serial"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeterListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meters/{meter_id}": {
      "get": {
        "summary": "Get Meter",
        "description": "Full detail for a single meter: nameplate + hierarchy + location.\n\nHierarchy and location are sourced from the bulk export (see\nPROTOCOL.md) since the portal has no clean single-meter detail\nendpoint with hierarchy included.",
        "operationId": "get_meter_api_v1_meters__meter_id__get",
        "parameters": [
          {
            "name": "meter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeterDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meters/{meter_id}/consumption": {
      "get": {
        "summary": "Get Meter Consumption",
        "description": "30-minute interval consumption history for a single meter.",
        "operationId": "get_meter_consumption_api_v1_meters__meter_id__consumption_get",
        "parameters": [
          {
            "name": "meter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Meter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConsumptionReading"
                  },
                  "title": "Response Get Meter Consumption Api V1 Meters  Meter Id  Consumption Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transformers": {
      "get": {
        "summary": "List Transformers",
        "description": "List distribution transformers (paginated).",
        "operationId": "list_transformers_api_v1_transformers_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransformerListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ConsumptionReading": {
        "properties": {
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "kwh": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kwh"
          },
          "kvah": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kvah"
          },
          "volt_r": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volt R"
          }
        },
        "type": "object",
        "required": [
          "timestamp"
        ],
        "title": "ConsumptionReading"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HierarchyNode": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          }
        },
        "type": "object",
        "title": "HierarchyNode"
      },
      "MeterDetail": {
        "properties": {
          "meter_id": {
            "type": "string",
            "title": "Meter Id"
          },
          "serial_no": {
            "type": "string",
            "title": "Serial No"
          },
          "make": {
            "type": "string",
            "title": "Make"
          },
          "phase_type": {
            "type": "string",
            "title": "Phase Type"
          },
          "install_status": {
            "type": "string",
            "title": "Install Status"
          },
          "install_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Install Type"
          },
          "dt_code": {
            "type": "string",
            "title": "Dt Code"
          },
          "hierarchy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MeterHierarchy"
              },
              {
                "type": "null"
              }
            ]
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MeterLocation"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "meter_id",
          "serial_no",
          "make",
          "phase_type",
          "install_status",
          "dt_code"
        ],
        "title": "MeterDetail"
      },
      "MeterHierarchy": {
        "properties": {
          "zone": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HierarchyNode"
              },
              {
                "type": "null"
              }
            ]
          },
          "circle": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HierarchyNode"
              },
              {
                "type": "null"
              }
            ]
          },
          "division": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HierarchyNode"
              },
              {
                "type": "null"
              }
            ]
          },
          "subdivision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HierarchyNode"
              },
              {
                "type": "null"
              }
            ]
          },
          "substation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HierarchyNode"
              },
              {
                "type": "null"
              }
            ]
          },
          "feeder": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HierarchyNode"
              },
              {
                "type": "null"
              }
            ]
          },
          "dt": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HierarchyNode"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "MeterHierarchy"
      },
      "MeterListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/MeterSummary"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "pageSize": {
            "type": "integer",
            "title": "Pagesize"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total",
          "page",
          "pageSize"
        ],
        "title": "MeterListResponse"
      },
      "MeterLocation": {
        "properties": {
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          }
        },
        "type": "object",
        "title": "MeterLocation"
      },
      "MeterSummary": {
        "properties": {
          "meter_id": {
            "type": "string",
            "title": "Meter Id"
          },
          "serial_no": {
            "type": "string",
            "title": "Serial No"
          },
          "make": {
            "type": "string",
            "title": "Make"
          },
          "phase_type": {
            "type": "string",
            "title": "Phase Type"
          },
          "install_status": {
            "type": "string",
            "title": "Install Status"
          },
          "dt_code": {
            "type": "string",
            "title": "Dt Code"
          }
        },
        "type": "object",
        "required": [
          "meter_id",
          "serial_no",
          "make",
          "phase_type",
          "install_status",
          "dt_code"
        ],
        "title": "MeterSummary",
        "description": "A single row as returned by the meters list/search endpoint."
      },
      "TransformerListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/TransformerSummary"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total",
          "page",
          "page_size"
        ],
        "title": "TransformerListResponse"
      },
      "TransformerSummary": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "feeder_code": {
            "type": "string",
            "title": "Feeder Code"
          },
          "capacity_kva": {
            "type": "number",
            "title": "Capacity Kva"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "feeder_code",
          "capacity_kva"
        ],
        "title": "TransformerSummary"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}